# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Agent Chat API - OpenClaw + Hermes features
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Session {
  id: string;
  title: string;
  model: string;
  channel: string;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
  /** @nullable */
  lastMessage?: string | null;
}

export interface SessionInput {
  title: string;
  model?: string;
  systemPrompt?: string;
}

export interface SessionUpdate {
  title?: string;
  systemPrompt?: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
  system: 'system',
  tool: 'tool',
} as const;

export interface Message {
  id: string;
  sessionId: string;
  role: MessageRole;
  content: string;
  /** @nullable */
  toolName?: string | null;
  /** @nullable */
  toolResult?: string | null;
  /** @nullable */
  tokens?: number | null;
  createdAt: string;
}

export interface MessageInput {
  content: string;
  model?: string;
  stream?: boolean;
}

export interface MemoryEntry {
  id: string;
  key: string;
  value: string;
  category: string;
  createdAt: string;
  updatedAt: string;
}

export interface MemoryInput {
  key: string;
  value: string;
  category: string;
}

export interface Skill {
  id: string;
  name: string;
  description: string;
  content: string;
  usageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface SkillInput {
  name: string;
  description: string;
  content: string;
}

export type ToolRunInputParams = { [key: string]: unknown };

export interface ToolRunInput {
  tool: string;
  params: ToolRunInputParams;
}

export interface ToolResult {
  success: boolean;
  result: string;
  /** @nullable */
  error?: string | null;
}

export interface AgentStats {
  totalSessions: number;
  totalMessages: number;
  totalMemoryEntries: number;
  totalSkills: number;
  telegramConnected: boolean;
  uptime: number;
}