# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Agent Chat API - OpenClaw + Hermes features
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /memory:
    get:
      operationId: listMemory
      tags: [memory]
      summary: Get all memory entries
      responses:
        "200":
          description: Memory entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryEntry"
    post:
      operationId: createMemory
      tags: [memory]
      summary: Create a memory entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryEntry"
  /memory/{id}:
    delete:
      operationId: deleteMemory
      tags: [memory]
      summary: Delete a memory entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /skills:
    get:
      operationId: listSkills
      tags: [skills]
      summary: List all agent skills
      responses:
        "200":
          description: Skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
    post:
      operationId: createSkill
      tags: [skills]
      summary: Create a new skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
  /skills/{id}:
    delete:
      operationId: deleteSkill
      tags: [skills]
      summary: Delete a skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    MemoryEntry:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
        value:
          type: string
        category:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - key
        - value
        - category
        - createdAt
        - updatedAt
    MemoryInput:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
        category:
          type: string
      required:
        - key
        - value
        - category
    Skill:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        content:
          type: string
        usageCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - description
        - content
        - usageCount
        - createdAt
        - updatedAt
    SkillInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        content:
          type: string
      required:
        - name
        - description
        - content
...