# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Agent Chat API - OpenClaw + Hermes features
 * OpenAPI spec version: 0.1.0
 */

export interface Skill {
  id: string;
  name: string;
  description: string;
  content: string;
  usageCount: number;
  createdAt: string;
  updatedAt: string;
}