# lib/api-zod/src/generated/types/skillInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Agent Chat API - OpenClaw + Hermes features
 * OpenAPI spec version: 0.1.0
 */

export interface SkillInput {
  name: string;
  description: string;
  content: string;
}